Moved the second area to the right, the third to the left of the screen.
The main reason for swapping the two is that now,
cycling through the windows is a lot more intuitive.

A possible drawback of this is that when the second
area is empty, the main area will be on the right side.
This could be changed in the future by checking if
nsecond == 0, but only if this doesn't make even less
intuitive.