Changed the bluetooth keybinding to MOD+b, added keybindings for turning on the displays.
One small issue is that, after turning displays
off and on again, the positioning is correct,
but the order in which dwm has them stored is
wrong (MOD+. behaves unexpectetly).
The current workaround is to first turn on the
internal display, then the second external one,
since this seems to work as expected.